backend/pyproject.toml
[project]
name = "open-learning-backend"
version = "0.1.0"
description = "OpenLearning AI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "langchain-anthropic>=0.3",
    "langchain-core>=0.3",
    "langgraph>=0.2",
    "langsmith>=0.1",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyyaml>=6.0",
    "sse-starlette>=2.2",
    "python-dotenv>=1.0",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.20",
    "langgraph-checkpoint-sqlite>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.9",
]

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B008", "RUF012"]

[tool.ruff.lint.per-file-ignores]
"app/main.py" = ["E402"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
